server/requirements.txt
python-dotenv
langchain
langchain-community
tavily-python
Pyrebase
pinecone-client
langchain_ollama